test(auth): migrate AuthService test to integration
